+++ src/Container/SimpleContainer.php
@@ -8,6 +8,7 @@ use Closure;
 use Psr\Container\ContainerInterface;
 use Throwable;
 use Yiisoft\Test\Support\Container\Exception\NotFoundException;
+
 use function array_key_exists;
 
 final class SimpleContainer implements ContainerInterface
@@ -30,7 +31,7 @@ final class SimpleContainer implements ContainerInterface
     public function get($id)
     {
         if (!array_key_exists($id, $this->definitions)) {
-            $this->definitions[$id] = ($this->factory)($id);
+            $this->definitions[$id] = ($this->factory)($id, $this);
         }
         return $this->definitions[$id];
     }