+++ tests/Container/SimpleContainerTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Test\Support\Tests\Container;
 
 use Closure;
+use Psr\Container\ContainerInterface;
 use Yiisoft\Test\Support\Container\SimpleContainer;
 
 final class SimpleContainerTest extends BaseContainerTest
@@ -28,6 +29,13 @@ final class SimpleContainerTest extends BaseContainerTest
         $container = $this->createContainer(['foo' => 'foo'], static fn ($id) => 'bar');
 
         $this->assertSame('foo', $container->get('foo'));
+    }
+
+    public function testClosureWithContainerInterface(): void
+    {
+        $container = $this->createContainer([], static fn (string $id, SimpleContainer $container) => $container);
+
+        $this->assertSame($container, $container->get(ContainerInterface::class));
     }
 
     protected function createContainer(array $definitions = [], Closure $closure = null): SimpleContainer